requirements.txt
annotated-types==0.7.0
anthropic==0.8.1
anyio==3.7.1
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cookiecutter==2.6.0
cycler==0.12.1
distro==1.9.0
Django==4.2.16
django-browser-reload==1.17.0
django-debug-toolbar==4.4.6
django-tailwind==3.6.0
djangorestframework==3.14.0
exceptiongroup==1.2.2
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
jiter==0.8.2
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
numpy==2.0.2
openai==1.12.0
packaging==24.2
pillow==10.4.0
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
text-unidecode==1.3
tokenizers==0.21.0
tqdm==4.67.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.8.2
zipp==3.21.0
